Move FloatOrNone to gmaps_traitlets.
This removes a circular dependency between ipy_compat and traitlets. FloatOrNone was more than just a compatibility layer anyway, so it makes sense to move it to the custom traitlets module.
