+++ gmaps/plainmap.py
@@ -1,5 +1,5 @@
 
-from ipy_compat import widgets
+from _ipywidgets import widgets
 from _traitlets import List, Unicode
 
 import gmaps.gmaps_traitlets as gmaps_traitlets
